#22 rename: 인증 관련 패키지 세분화
