+++ src/main/java/org/com/itpple/spot/server/common/auth/userDetails/UserDetailsCustomService.java
@@ -1,4 +1,4 @@
-package org.com.itpple.spot.server.common.jwt;
+package org.com.itpple.spot.server.common.auth.userDetails;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
